CORE-3828: Speed-up execution of concurrent E2E test
With 100 denies and 100 allows it takes around 20 seconds for the whole scenario.

When creating permissions, we save time by not verifying it.
Also adding permissions to a role in bulk.

Can improve speed even further if during `createPermission` HTTP client and proxy is re-used. Creation a new instance every time is somewhat costly.