NOTICK Change Gateway to always route same initiatorHello to same LinkManager (#917)
Before this change we would occasionally see the MAC of the InitiatorHandshake to fail authentication (when testing a HA deployment). This was caused by the initiatorHello message being replayed and routed to a different LinkManager. Meaning we end up with two different responderHello messages. This causes the mac check to fail in the InitiatorHandshake. This change uses the session identifier as the key to ensure replays fall into the same partition.