CORE-4075 Make ReplaySchedulerTest deterministic (#927)
Mock the ScheduledExecutorService in the same way as in the HeartbeatManager. To do this we pass a factory into ReplayScheduler. The default factory argument which creates a new ScheduledExecutorService is used in the LinkManager. Under test conditions we pass in a mocked component. I also updated HeartbeatManager to also use a factory so that the ScheduledExecutorService can be used as a resource. The executor service will shut down if the dominoTile stops (this was the behaviour prior to commit 4b1112e).