NOTICK - Remove flakiness from test for outbound message handler (#1011)
This replaces the scheduled executor service in the outbound message handler test with the mock one to remove flakiness from the test. The mock time facilities were also extracted in the testing-utilities library. I had to replace mockito-kotlin with mockito-core, because the former lacks the necessary OSGi attributes causing resolution issues.