CORE-4154 Move serialization into `FlowSessionImpl`
Move serialization and deserialization of sent and received payloads
into `FlowSessionImpl`. This means the pipeline handlers that related to
the session behaviour now pass around byte arrays rather than
serializing or deserializing them into objects themselves.

This is done to open the possibility for different types of
serialization methods to be used in the future.