CORE-4292 Improved CPI uploading and error handling (#295)
* CORE-4292 Improved CPI uploading and error handling

Renamed UploadStatus to ChunkReceived for clarity.

Added ChunkKey because we need a unique identifier for each
ChunkReceived message.

Reduce enumeration to ok, failed, or in progress.

Add the part number back in.  We need this because we could have multiple
publishers and subscribers so the "compacted queue" of any subscriber needs
to be the complete message set.  An out-of-order receipt of previous
UploadStatus messages could lead to a perpetual "in progress" state.

* CORE-4292 Remove explicit status from Chunk acks

Renamed `ChunkReceived` to `ChunkAck` which is consistent with other
messages.

We can infer the state (ok, in progress, failed) from the content
of the messages.

We also remove redundant data from the `ChunkAck` since
we send `(ChunkAckKey, ChunkAck)` pairs which contains the same
information.

Simplify unit tests further.

* Convert back to UploadStatus, sort of