CORE-3456: Ensure that call to `GetProtocolVersion` is exempt from permissions check. (#1035)
It appears that none of the existing integration tests been catching this situation.
Latest set of RESTful changes meant that calls to `GET testEntity/getprotocolversion` cannot be distinguished from `GET testEntity/:id`. The former should be an exempt from permissions check, whereas the later should not.

The integration tests been added as necessary.