CORE-4145 Check for `@InitiatingFlow` in `initiateFlow` (#1065)
Check that the current flow (the flow on top of the flow stack) is
annotated with `@InitiatingFlow` when calling
`FlowSession.initiateFlow`.

If the flow is not annotated, then throw an error.

Otherwise, add the session id to the current `FlowStackItem` and create
the session.

The session will be late initiated when calling send/receive, however,
the session id itself will be associated at the time of calling
`initiateFlow`. This ensures that the flow is associated to the flow
calling `initaiteFlow` rather than a possible subflow, if the created
session is not used before being passed to a subflow.