Turn off exponential backoff.
Split up the Replay calculator into an abstract class,
ExponentialBackoffReplayCalculator and ConstantReplayCalculator.
For the moment only ConstantReplayCalculator is used in the Link
Manager. As a part of CORE-4572 we will renable expoential backoff.
Also respond to review comment: if total weight is zero instead of
weighting everything with 1 just select a random session (this has
no functional affect, just makes the code cleaner).