CORE-3727: Unify P2P Tools (WIP) (#1067)
This PR is unifying the three P2P tools we have now into a single tool (that can run once and produce groups, members, locally hosted identities, key-pairs and configurations).

There are two ways to run the tools, via apply and a file that contains all the details or via different sub-commands to perform each operation (this can also be loaded from a file using the @file option). More details are in the README file.