CORE-4570 Send Heartbeats Eagerly from the Link Manager (#1113)
Before we only started to send Heartbeats when we sent the first data message on a specific session. This made sense when we only had one session per counterparty. Now (after CORE-2732) we have more than 1, we might not send a data message for a while (if we only want to send one message for example). In this scenario the session might time out due to never sending a data message. This could lead to idle sessions repeatedly timing out in low traffic situations.