CORE-4146 Close sessions when completing subFlow
Close any sessions related to an initiating subFlow when successfully
completing the subFlow.

The close occurs in `SubFlowFinishedRequestHandler`, causing the flow to
wait for the closes to be acknowledged or instantly waking up the flow
if the sessions are already closed or there are none to close in the
first place.

An extra suspension has been added to the end of `FlowFiberImpl.run` to
ensure that close messages / acknowledgements are always sent. This is
needed because the `FlowFinishedRequestHandler` removes the checkpoint
and prevents the `FlowGlobalPostProcessor` from accessing any sessions
and sending messages. This needs to be reviewed in the future.

The `ReceiveRequestHandler`, `SendAndReceiveRequestHandler` and
`CloseSessionsRequestHandler` now contain logic to determine if the flow
should instantly wake up (by scheduling a `Wakeup` event).
- For receives, it checks if messages are received for each session
being waited for.
- For closes, it checks if the sessions are already in the states of
`CLOSED` or `WAIT_FOR_FINAL_ACK`.

Also includes a few identity corrections to initiated, initiating and
counterparty identities on session events and states.