CORE-4156 Check payloads in `FlowSession` receive methods
In `FlowSession.receive` and `FlowSession.sendAndReceive`, check the
received payload type.

The wrong payload could be send back to the
receiving session, and the AMQP deserialization will not inherently fail
by itself.

Therefore, a check has been added to ensure the received type is the
same type expected by the session.