NOTICK - Make p2p e2e test assertions on markers a bit more lenient (#1141)
I noticed some flaky failures of the p2p end-to-end tests after the recent changes for TTL. The reason appears to be that we had assertions on the exact number of markers, which could differ depending on the execution (e.g. if a message is replayed, under race conditions an acknowledgement marker might be written more than once). I adjusted these assertions to just check that at least one marker of the expected type has been written for every message.