CORE-4540 Push all close logic into handlers (#1160)
Remove session close state checking in `FlowSessionImpl` and rely on the
logic in `CloseSessionsRequestHandler` instead.

Sessions are not removed from the current flow stack item now, which
means when a subflow or flow finishes it will implicitly clean up the
sessions due to popping the flow stack item off.

This means sessions that even if a subflow's sessions are already
closed, a suspension will still occur but might instantly resume if
they're already closed. Similar behaviour is done at the end of a flow.

The `initiated` state of `FlowSessionImpl` is still kept as it's
required for the lazy session init, which the handlers cannot help with,
as reaching the handlers requires a suspension in the first place.