CORE-4153 A subFlow failing sends errors to sessions
When a subFlow fails errors are now sent to sessions that it initiated
(initiated flows treat the initiating session as an initiated session).

The subFlow finished and session closing logic has also been updated to
handle errored sessions.

When erroring sessions via `SubFlowFailedRequestHandler`, `CLOSED` and
`ERROR`'d sessions are filtered out, meaning that they do not receive
session error events. A `Wakeup` event is always scheduled as the flow
is meant to move on instantly after sending the session error events.

When closing sessions via `CloseSessionsRequestHandler` and
`SubFlowFinishedRequestHandler`, `ERROR`'d sessions are filtered out,
meaning that they do not receive session close events and are not waited
for when resuming the flow. This means that if all sessions are closed
or errored (or a mixture), then the flow will schedule a wakeup event as
there are no session close events to receive.

The "waiting for" handlers have also been updated to handle the
potential errors from peer sessions.

When waiting for confirmation of a session init, if an error is
received, the flow resumes with an error.

When waiting for confirmation of session closes, the flow will only
resume if all sessions have moved to the status of `CLOSED` or `ERROR`.
If there are any errors, then the flow resumes with an error.

When waiting to receive session data, the flow will only resume if all
the sessions have either received their next session data event or have
moved to the status of `ERROR`. If there are any errors, then the flow
resumes with an error.

`CordaRuntimeException`s have currently been used to report any session
errors to flows.

This task has not made changes to calling `send`, `sendAndReceive` or
`receive` on errored or closed sessions.