CORE-4095 Correct bnd scan for entities (#348)
Since we're serializing the entities using AMQP, we need to also mark
them as `@CordaSerializable`.  If we don't, the bnd scanner will produce
a list of every `@Entity` in a cpk, rather than the subset of ones
that we can actually use to persist from flow (worker) to db (processor).