CORE-3779 Investigate using pub sub subscription in gateway (#1161)
Restructure the OutboundMessageHandler in the gateway to use a pubsub subscription instead of EventLog.
Changed how the response is handled (as the current method no longer made sense when using a pubsub subscription):

Changed PubSubSubscription so that messages can be processed asynchronously. We also no longer commit the offset. To process records asynchronously onNext (inside PubSubSubscription) returns a future. The next batch of records will only be processed after the future for each record has completed.