CORE-3350 Remove components and supporting libs that publish config straight to Kafka (#1264)
Removed the following Libraries:
:components:configuration:configuration-publish-service
:components:configuration:configuration-publish-service-impl
:libs:configuration:configuration-publish
:libs:configuration:configuration-publish-impl

These publish config directly to kafka bypassing the database. Refactored the kafka-setup tool + p2p tests to use a publisher to publish config instead. Preserved most of the functionality in the kafka setup tool with the exception of some logic which strips out strings from the component version. As this is no longer needed.

I have tested corda-kafka-setup, deploying a local Kafka cluster in docker. I can publish config to the Kafka cluster tested it against an older version of the corda-kafka-setup tool.