CORE-4783 Entity processor error handling
The persistence service / entity processor now provides richer error
information to the flow worker.

Internally, we're returning the response messages everywhere so that
we can just batch them up and let `onNext` publish them for us.

We have added several categories of error that may or not may not be
recoverable via the flow-worker retrying the command and possibly
connecting to a different db worker.

A fatal error in one db worker is not recoverable (though possibly retrying
with a different db worker *may* work).