Update README.md
Reworded the readme to make it a bit more welcoming, signpost people to Corda 4 if they're looking for the current version, and to clarify the contribution process.   NOTE: I think we need to provide a way to contact us (specific email address?) and also need to link to a file equivalent to this https://github.com/corda/corda/blob/release/os/4.10/CONTRIBUTING.md  to ensure that third party PRs are appropriately licensed.