CORE-4805 Discard orphaned session events
When a flow finishes session events could still arive for the flow due
to a race condition between the flow mapper and flow processor.

We now discard session events that were received for a flow that does
not exist. In most cases, this will be receiving a session event for a
flow that has finished or failed already.