CORE-4877 deleteById/findAll persistence methods
Moved Avro messages to own package.

Added DeleteById and FindAll persistence methods.

Placeholder code to check that we do not attempt to return a
result set larger than a Kafka packet.

Use the `failureResponse` method everywhere and also log any caught
exceptions, as errors if fatal, or warnings if not.  We want
to record any callstack or exception information particularly in the
case of a fatal error so we can investigate further.

Tidied up unit tests for readability.  Fixed a couple of bugs.

Make one of the test bundles have a composite key.

Replace usage of `Instant` with `Clock`. At component level, just create
a local `UTCClock` but then pass it in to places where it is _used_ so
we can substitute later if necessary.