CORE-4851 Discard orphaned `Wakeup` events (#1336)
If a flow does not exist and a `Wakeup` event is received, then discard
the event. It is highly likely that a `Wakeup` received this way was due
to a flow finishing but still having a `Wakeup` event in the message
queue, that was either scheduled by the flow or by the flow heartbeating
mechanism.