CORE-3972 boot config updates (#1310)
Updates to how boot configuration is handled in the runtime os. cmd line args to workers are processed into a boot config which must conform to a schema defined in the runtime os project. Other config keys are validated in the rpc worker and db workers and defaults are applied from the json schemas when writing to kafka. Config read from the topics via the config read service are merged with the boot config explicitly via the ConfigMerger api instead of typesafeconfig fallback mechanisms.