CORE-4802 Unbalanced sends/receives/closes
Verify that receiving a `SessionData` event when a flow is expecting a
`SessionClose` event resumes the flow with an error.

Verify that receiving a `SessionClose` event when a flow is expecting a
`SessionData` event resumes the flow with an error.

This required altering the `SessionDataWaitingForHandler` to error the
flow if a `SessionClose` is received. Note, it only errors the flow once
all the sessions have either received data events, errored or a closing.