CORE-4998 Cleanup flow mapper in-memory state
Clean up flow mapper in-memory state of flow starts and sessions.

When a flow finishes of fails, clean up its state.

When a session becomes `CLOSED` or `ERROR`ed, clean up its state.

State is cleaned up by sending a `ScheduleCleanup` record to the flow
mapper, which cleans up the in-memory state after a period of time.

`FlowCheckpointImpl` now _soft deletes_ the Avro `Checkpoint` so that
sessions can still be accessed after marking the `Checkpoint` for
deletion. This is needed if a session goes to `CLOSED` and the flow
finishes all off of a single event, otherwise the `ScheduleCleanup`
record cannot be sent for the session.

This commit also changes the flow mapper's key solely to a `String`. The
string still represents a `FlowKey` (for flow starts) or a session id
(for a session). The change calls `toString` on `FlowKey` rather than
passing in the `FlowKey` directly and letting the state and event
pattern do the conversion to a `String`. This was required to fix a bug
where the partition a state/event was read from was not the same
partition the new state was written to, which indicated a change in the
generated hash of the key. Calling `toString` made the output
consistent and preventing changing the partition.