CORE-3153 - Add integration with crypto ops client in the gateway (#1480)
Adding integration with the `CryptoOpsClient` in the gateway to delegate TLS signing to. The existing integration with our stub crypto processor is kept in place so that end-to-end testing can be performed in our existing setup and will be removed once integration with MGM/dynamic networks has been completed and tested succesfully. An option has been added to switch between stubs or the real crypto processor in the Gateway worker. The default has been set up to use the stubs, so that our end-to-end tests continue working without further changes needed.