CORE-3150 - Add a switch for integrating with membership/crypto components (#1489)
Adding a switch that can be used as a toggle when integrating with the membership/crypto components in the link manager, so that we can temporarily maintain both versions until we have confirmed integration is working successfully with dynamic networks. Adding just this switch separately, so that we can create separate PRs for each component and deliver things a bit more incrementally.