CORE-4880 Persist cpk Liquibase scripts
Persist any Liquibase script found in a cpk in to the database.

Perform a cursory check that we can parse the XML, find the root element
and that it has an `xmlns` attribute set.

Write liquibase entities in same 'cpi' transaction

Collect entities and pass them into the
"main" transaction that writes _all_ CPI entities to the database
so that we have a single transaction and single rollback.

Added unit tests.

Moved InMemoryZipFile to testing project.

Renamed classes in chunking library to be more descriptive.