CORE-5134 Add Helm Chart for Link Manager (#1509)
Changed the Link Manager command line arguments to be compatible with the Helm Charts i.e. the same as the common worker:
1. Removed the option `--kafka-servers` instead we accept a list of arguments prefixed with -m. So for example instead of `--kafka-servers=localhost:9092` use `-mbootstrap.servers=localhost:9092`. By passing in multiple options TLS can be turned on for Kafka.
2. Changed `--topic-prefix` to `--topicPrefix`.

Currently the link manager does not use the common worker framework (this will be done as a part of CORE-4424). For this reason there is no health port exposed. Hence the Link Manager will pod will be ready even if it has invalid config.