Remove scheme check from gateway server (#1566)
Removing the scheme check from the gateway server, because checks for TLS are already happening at a level above (so this is not really needed to check that it's a TLS connection) and it's causing issues with clients that might be sending an absolute path as part of the HTTP request (`POST /gateway/send HTTP/1.1` instead of `POST https://www.alice.net:3001/gateway/send HTTP/1.1`) which is something reasonable. I also adjusted one of the tests we had for invalid requests as before it was falling into this case (I believe by accident).