CORE-3153 Replace Stub Crypto Processor With Real Implementation in LinkManager (#1560)
Integrate the CryptoClient with the LinkManager. So it can be used for signing with the Session Private Key when generating session negotiation messages, during Diffie-Hellman key exchange. Like the membership integrations by default we use the stub component. Using the real `CrypoClient` can be turned on using the option `--without-stubs`.