CORE-5451 Expose member lookup based on session key (#1573)
Extends the membership group reader implementation to expose a lookup of member info based on session initiation key. This works in the same way as the ledger key lookup except uses a single key instead of a list. Key hash must be set in the member context for faster lookups. New lookup called lookupBySessionKey to avoid duplicate function signature.