CORE-5459 Find nearest initiating flow when initiating session
We check whether a session is created inside or outside an initiating
flow. However, when passing a session created in an initiating flow that
has not already been interacted with/initiated into a subFlow, then we
were trying to find the matching responder of the subFlow instead of the
initiating flow. This is due to the lazy initialisation of sessions.

This caused an error as there was no matching responder flow for the
subFlow.

We now correctly find the nearest initiating flow in the flow stack when
initiating a flow for the first time.