CORE-5518 Verify Quasar instrumentation in flow worker
Setting `-Dco.paralleluniverse.fibers.verifyInstrumentation=true`
enables Quasars instrumentation verification to check for correct use of
`@Suspendable`s.

For Kafka deployments, the default value is `false` with an entry in
`debug.yaml` that sets it to `true`.

For the combined worker, the run configuration have been updated to
include the property. Otherwise, it is up to the user to include the
property when running the jar.