CORE-3151 Replace membership group reader with the real implementation in the link manager (#1585)
Integrate the MembershipGroupReader with the LinkManager. By default the stub component is still used, the real Membership Group Reader can be turned on using the option --without-stubs.

The real MembershipGroupReader had a slightly different API to the one used by our stub component (`StubMembershipGroupReader`): it needs to know the looking up of member's holding identity, as different members can have a different view of the Membership group reader. To save time, because the stub code will soon be removed, I did not update the stub to behave in the same way. The stub ignores the looking up member's holding identity and always returns the same `MemberInfo` for a given `lookupIdentity`.