CORE-5305: Use of Websockets in HTTP RPC PoC (#1572)
This Proof Of Concept (PoC) work demonstrates use of Websocket protocol for receiving streaming data.
It shows how to integrate Websocket endpoints into HTTP RPC Worker and Javalin Webserver.

Websocket endpoints cannot be represented in OpenAPI and, therefore, excluded from Swagger.

In this PoC Websocket initiating endpoint is meant to be annotated with `HttpRpcWS` and `RpcOps` method definition has a special type `DuplexChannel` as a first argument which enables writing server side code and employ a two-way communication.

Recommending to start review with `TestHealthCheckAPIImpl.counterFeed()` which shows how data feed can be implemented. Then it should be possible to have a look at the new integration test `HttpRpcServerWebsocketTest`, which stands-up in-process web-server with Websocket endpoint.

Permission control is also enforced on all the Websocket endpoints using existing `HttpRpcSecurityManager`.

In addition to the integration it was also possible to perform Websocket call using external client - Postman.

Integration tests been created which prove that WebSockets are callable programmatically over HTTP and HTTPs.