CORE-5601 Unified external events
The sending on events to external workers is unified to simplify the handling of sending requests and receiving responses. This reduces the amount of code we have to write to add new functionality that requires communication with other workers.

As we are making a generic solution there will be limitations that each specific use-case requires, however a number of scenarios have already been identified and the solution _should_ cater for them all.

The generic code handles the sending of events to other workers, receiving `ExternalEventResponse`s from them and determining whether to resume a flow or not based on the responses.

## How to use

After these changes, to send an event/request to a worker and receive a response in a service running in a flow, the following must be done:

### Inside the flow worker

- Inject `ExternalEventExecutor` into the service that needs to send an event.
- Create an OSGi component that implements the `ExternalEventHandler` interface and make it discoverable via that interface.
- In the `ExternalEventHandler.suspending` implementation, use the `ExternalEventContext` passed in to set the `flowExternalEventContext` property of the created payload/event.
- Call `ExternalEventExecutor.execute` and pass it the created `ExternalEventHandler`.
- Once the response is received, the flow will resume with the `RESUME` type defined in `ExternalEventHandler`.

### Inside the external worker

- Define a `DurableProcessor` or `StateAndEventProcessor` that receives whatever Avro object that the processor wants to receive. The external event framework does not restrict this.
- Process the event, which ends in either a successful response or an error response.
- Use `ExternalEventResponseFactory` to create a successful response or error response as required.