CORE-4882: remove owner changelog (#470)
* CORE-4882: remove owner changelog

Having the changelog (liquibase migrations) in corda-api was a shortcut until we have liquibase migrations
being picked up from CPK files embedded in CPI files when virtual nodes were being created. This has now been
done, and to make that testable we had to remove a couple of changelogs to avoid migrations being applied
multiple times. It wasn't necessary to remove the owner changelog to get the tests to work and the previous PR
was already large, so now following through and removing the remaining owner changelog.

There are still changelogs in corda-api, which are for setting up parts of the schema which are unrelated to
specific CPIs.

* CORE-4882: remove incomingchatmessage creation changelog