CORE-5682: handle each CPK separately, remove relative import support and normalize paths (#1642)
Our original idea was to run all the separate CPK migrations in a CPI in one invocation of Liquibase,
so as to run everything in one transaction. That introduces a difficulty when a migration includes
another migration file, since it isn't clear what migration is required. One option is to use
relative paths, and we started to do that. However, that required all imports to either be marked
relative by the migration author or to use the full CPK name in the path, which involves baking
the CPK name into the migration. So, instead, we now invoke liquibase once per CPK, so we know
where to resolve include files when liquibase calls us back.

We had only implemented partial support for relative imports in liquibase migrations.
It was a significant piece of work to finish the support, and with the per-CPK migraiton change
we don't need to support relative imports, so for the sake of simplicity for the time being
let's remove support relative imports. If liquibase does call us with relativeTo, we throw
UnsupportedArgumentException.

This does involve making an assumption that provided callers do not specify "relativeToChangeLog"
liquibase will not call us with a relativeTo path. That appears to be the case by observation,
but we haven't got deep enough into understanding the liquibase codebase to be convinced our
assumption always holds.

While experimenting with relativeTo support we discovered that internally liquibase normalises paths
by replacing runs of consecutive forward slashes with a single slash. That causes problems with
our "classloader://" prefix convention, which we've seen in connection to relative paths. In such
cases liquibase passes back its internal notion of the normalized path to us. Since we failed
to interpret the normalised prefix "classloader:/" as our prefix we passed it in when querying
the filesystem, which causes an exception on Windows where ':' is an illegal pathname component.
Being liberal in what we accept is good practice, so, even though we don't necessarily need the
normalisation of slashes we'll do so.

While touching the code:
 - improve test coverage including one that illustrate an ordering assumption we have to make with
   the liquibase callback interface.
 - refactor to use more Kotlin idioms so as to shorten and improve readability in places.
 - improve logging at both error and info levels