CORE-5751 Add TLS support to the Application Simulator (#1673)
Modified the AppSimulator to take -m options to work in the same way as the LinkManager, Gateway and the common worker framework. To set the kafka to connect to a broker on kafka-broker:1000 use -mbootstrap.servers=kafka-broker:1000. To enable TLS add the following options: -msecurity.protocol=SSL, -mssl.truststore.location=/certs/ca.crt and -mssl.truststore.type=PEM. Where /certs/ca.crt is the trust store.