CORE-5674 Retry topic creation
There doesn't appear to be a reliable way to check for topics that are
still pending deletion so, instead, we'll just retry the creation if it
fails due to a topic already existing. Testing locally we seem to go
round the loop 2 or 3 times before everything then deletes successfully.

As recommended by Confluent, we're also dropping the message retention
time before deleting the topics.