CORE-5408: Improvements to the modules packaging
In particular, dependency: `api project(':libs:packaging:packaging-core')` is a problem as intention is to keep HTTP endpoints as lean as possible in terms of dependencies.

This might be important, if Java HTTP RPC Client is created against the interface it will not need to pull an extensive graph of transitive dependencies.