CORE-5896 Lazily initialise MGM group policy for data coming from the DB (#1718)
Group policy data for the MGMs group policy is provided at onboarding/registration time. The group policy file needs to be accessible earlier to allow for registration protocol selection. This change updates how persisted properties are set so that they are lazily initialized so they aren't queried for until they are needed.