CORE-5338: support named queries in the persistence service (#1684)
Add integration tests. This involves completing fixture work to make
sure that each test case runs independently. Since in memory testing database
does not pay attention to schema we have to ensure it is keyed differently
for each test.

Rework the Dog schema to allow the owner column to be nullable and rework some
of the test data to have a null value so we can test for null lookups.