CORE-5913: Change schema version type for `ConfigRPCOps`
It is important not to leak internal `Version` type to public API
such that we could evolve them separately.

In theory this should be a non-breaking change from external API
callers of view. The only visible effect in OpenAPI should be the name of the type has changed from `Version` to `ConfigSchemaVersion`.