CORE-5886 De-duplicate VirtualNodeInfo in REST API
Remove the duplicate `VirtualNodeInfo` object, and reuse the existing
one in the REST API.

Rename a couple of REST objects such that they don't start with an
action-verb.

Change `cpiFileChecksum` to return the 12 char version consistent with
the returned value in the other REST api call, and add
`cpiFileFullChecksum` to return the full checksum (not sure if it
will ever be used though).

Also add HoldingIdentity endpoint type

Static network test is built with no knowledge of core corda types, so
we need to map internal `HoldingIdentity` to one in the endpoint
rpc-layer instead.

This also decouples the REST layer from the rest of the code base a
little bit more.