CORE-5898 Fix Domino Logic (#1733)
Check the internalState instead of isRunning under `handleChildStarted`. 

When we do `coordinator.updateStatus(it)` the status does not get updated straight away, the LifecycleCoordinator posts a `StatusChangeEvent` which eventually gets processed and updates the `status` (see `LifecycleCoordinatorImpl`). Thus if we stop the Tile in one event `isRunning` might still be true in the next event. This can happen if the next event gets added to the queue before the `StatusChangeEvent`.

This caused the gateway to fail to start sometimes. To fix this we instead check the internalState. This gets atomically updated, so will be seen by subsequent events.