CORE-5948: Allow pipeline state modification after deletion of the flow state (#1747)
The flow checkpoint will get marked as deleted when the flow completes. If the flow was in the retrying state though, it will then attempt to mark the retry as successful, which causes a failure as the checkpoint is marked as deleted.

With the new split between pipeline and flow state, updating the pipeline state becomes irrelevant once the checkpoint has been marked as deleted anyway, so allowing this to happen does no harm. This might also be relevant if some transient error were to occur after the flow has been marked for deletion, so it is safer to take this approach than to check before access every time.

* CORE-5948: Allow the checkpoint pipeline state to be updated even if the flow state is marked for deletion

* CORE-5948: Remove superfluous isDeleted check when running the global post processor

* CORE-5948: Update the flow integration tests to cover the issue