CORE-5574 Add P2P Processors To Combined Worker (#1745)
Add the Link Manager and Gateway to the combined worker. Change the LinkManager and Gateway to not to instantiate Stub Components when these are unused. This change is necessary for the ready end point to become ready as previously these components would always have Lifecycle  status `DOWN`. The HealthMonitor checks every Lifecycle is `UP` before considering the worker ready.